Move implementation to separate file
