+++ pkg/git/remote_git_resolver_test.go
@@ -3,7 +3,7 @@ package git
 import (
 	"testing"
 
-	"github.com/go-git/go-git/v5/plumbing/transport/http"
+	"github.com/go-git/go-git/v6/plumbing/transport/http"
 	"github.com/sclevine/spec"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
